# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: UAT Test Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /releases:
    get:
      operationId: listReleases
      tags: [releases]
      summary: List all releases
      responses:
        "200":
          description: List of releases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Release"
    post:
      operationId: createRelease
      tags: [releases]
      summary: Create a release
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReleaseBody"
      responses:
        "201":
          description: Created release
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Release"
  /releases/{id}:
    get:
      operationId: getRelease
      tags: [releases]
      summary: Get release by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Release details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReleaseDetail"
    patch:
      operationId: updateRelease
      tags: [releases]
      summary: Update release
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReleaseBody"
      responses:
        "200":
          description: Updated release
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Release"
...
  /test-scenarios:
    post:
      operationId: createTestScenario
      tags: [testCases]
      summary: Create a test scenario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTestScenarioBody"
      responses:
        "201":
          description: Created test scenario
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestScenario"
  /test-scenarios/{id}:
    get:
      operationId: getTestScenario
      tags: [testCases]
      summary: Get test scenario by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Test scenario with cases
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestScenarioDetail"
  /test-cases:
    get:
      operationId: listTestCases
      tags: [testCases]
      summary: List test cases
      parameters:
        - name: testPlanId
          in: query
          required: false
          schema:
            type: integer
        - name: scenarioId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: priority
          in: query
          required: false
          schema:
            type: string
        - name: assigneeId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of test cases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TestCase"
    post:
      operationId: createTestCase
      tags: [testCases]
      summary: Create a test case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTestCaseBody"
      responses:
        "201":
          description: Created test case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestCase"
  /test-cases/{id}:
    get:
      operationId: getTestCase
      tags: [testCases]
      summary: Get test case by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Test case details with steps
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestCaseDetail"
    patch:
      operationId: updateTestCase
      tags: [testCases]
      summary: Update test case
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTestCaseBody"
      responses:
        "200":
          description: Updated test case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestCase"
...
components:
  schemas:
...
    Release:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        status:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        freezeDate:
          type: ["string", "null"]
          format: date-time
        goLiveDate:
          type: ["string", "null"]
          format: date-time
        ownerId:
          type: ["integer", "null"]
        ownerName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, name, description, status, startDate, endDate, createdAt]
    ReleaseDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        status:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        freezeDate:
          type: ["string", "null"]
          format: date-time
        goLiveDate:
          type: ["string", "null"]
          format: date-time
        ownerId:
          type: ["integer", "null"]
        ownerName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        testCycles:
          type: array
          items:
            $ref: "#/components/schemas/TestCycle"
        stats:
          $ref: "#/components/schemas/ReleaseStats"
      required: [id, name, description, status, startDate, endDate, createdAt, testCycles, stats]
    ReleaseStats:
      type: object
      properties:
        totalCycles:
          type: integer
        totalTestCases:
          type: integer
        passed:
          type: integer
        failed:
          type: integer
        blocked:
          type: integer
        notStarted:
          type: integer
        passRate:
          type: number
        openDefects:
          type: integer
      required: [totalCycles, totalTestCases, passed, failed, blocked, notStarted, passRate, openDefects]
    CreateReleaseBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        freezeDate:
          type: ["string", "null"]
          format: date-time
        goLiveDate:
          type: ["string", "null"]
          format: date-time
        ownerId:
          type: ["integer", "null"]
      required: [name, description, startDate, endDate]
    UpdateReleaseBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        freezeDate:
          type: ["string", "null"]
          format: date-time
        goLiveDate:
          type: ["string", "null"]
          format: date-time
        ownerId:
          type: ["integer", "null"]
    TestCycle:
      type: object
      properties:
        id:
          type: integer
        releaseId:
          type: integer
        name:
          type: string
        type:
          type: string
        status:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        ownerId:
          type: ["integer", "null"]
        ownerName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, releaseId, name, type, status, startDate, endDate, createdAt]
...
    TestScenario:
      type: object
      properties:
        id:
          type: integer
        testPlanId:
          type: integer
        title:
          type: string
        description:
          type: string
        businessProcess:
          type: string
        priority:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, testPlanId, title, description, businessProcess, priority, createdAt]
    TestScenarioDetail:
      type: object
      properties:
        id:
          type: integer
        testPlanId:
          type: integer
        title:
          type: string
        description:
          type: string
        businessProcess:
          type: string
        priority:
          type: string
        createdAt:
          type: string
          format: date-time
        testCases:
          type: array
          items:
            $ref: "#/components/schemas/TestCase"
      required: [id, testPlanId, title, description, businessProcess, priority, createdAt, testCases]
    CreateTestScenarioBody:
      type: object
      properties:
        testPlanId:
          type: integer
        title:
          type: string
        description:
          type: string
        businessProcess:
          type: string
        priority:
          type: string
      required: [testPlanId, title, description, businessProcess]
    TestCase:
      type: object
      properties:
        id:
          type: integer
        scenarioId:
          type: integer
        title:
          type: string
        description:
          type: string
        preconditions:
          type: string
        priority:
          type: string
        complexity:
          type: string
        status:
          type: string
        assigneeId:
          type: ["integer", "null"]
        assigneeName:
          type: ["string", "null"]
        estimatedMinutes:
          type: ["integer", "null"]
        lastExecutionStatus:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, scenarioId, title, description, priority, complexity, status, createdAt]
    TestCaseDetail:
      type: object
      properties:
        id:
          type: integer
        scenarioId:
          type: integer
        title:
          type: string
        description:
          type: string
        preconditions:
          type: string
        postconditions:
          type: string
        priority:
          type: string
        complexity:
          type: string
        status:
          type: string
        assigneeId:
          type: ["integer", "null"]
        assigneeName:
          type: ["string", "null"]
        estimatedMinutes:
          type: ["integer", "null"]
        lastExecutionStatus:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        steps:
          type: array
          items:
            $ref: "#/components/schemas/TestStep"
        executions:
          type: array
          items:
            $ref: "#/components/schemas/TestExecution"
      required: [id, scenarioId, title, description, priority, complexity, status, createdAt, steps, executions]
    TestStep:
      type: object
      properties:
        id:
          type: integer
        testCaseId:
          type: integer
        stepNumber:
          type: integer
        action:
          type: string
        expectedResult:
          type: string
        testData:
          type: string
      required: [id, testCaseId, stepNumber, action, expectedResult]
    CreateTestCaseBody:
      type: object
      properties:
        scenarioId:
          type: integer
        title:
          type: string
        description:
          type: string
        preconditions:
          type: string
        postconditions:
          type: string
        priority:
          type: string
        complexity:
          type: string
        assigneeId:
          type: ["integer", "null"]
        estimatedMinutes:
          type: ["integer", "null"]
        steps:
          type: array
          items:
            $ref: "#/components/schemas/CreateTestStepBody"
      required: [scenarioId, title, description, priority, complexity]
    CreateTestStepBody:
      type: object
      properties:
        stepNumber:
          type: integer
        action:
          type: string
        expectedResult:
          type: string
        testData:
          type: string
      required: [stepNumber, action, expectedResult]
    UpdateTestCaseBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        preconditions:
          type: string
        postconditions:
          type: string
        priority:
          type: string
        complexity:
          type: string
        status:
          type: string
        assigneeId:
          type: ["integer", "null"]
        estimatedMinutes:
          type: ["integer", "null"]
    TestExecution:
      type: object
      properties:
        id:
          type: integer
        testCaseId:
          type: integer
        testCycleId:
          type: integer
        executorId:
          type: ["integer", "null"]
        executorName:
          type: ["string", "null"]
        status:
          type: string
        result:
          type: string
        notes:
          type: string
        executedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        testCaseTitle:
          type: ["string", "null"]
      required: [id, testCaseId, testCycleId, status, result, createdAt]
...