# lib/api-zod/src/generated/types/release.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UAT Test Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Release {
  id: number;
  name: string;
  description: string;
  status: string;
  startDate: Date;
  endDate: Date;
  /** @nullable */
  freezeDate?: Date | null;
  /** @nullable */
  goLiveDate?: Date | null;
  /** @nullable */
  ownerId?: number | null;
  /** @nullable */
  ownerName?: string | null;
  createdAt: Date;
}